Fix issue 340. If the cursor colour is too close to a cell's background colour, use the cell's foreground colour instead. Before, it would have used the cursor text colour instead, which usually is the same as the default background colour. That resulted in the block cursor becoming invisible when it was on a single cell with a background colour close to the cursor colour.
Additionally, for a block cursor, if the cursor text colour then turns
out to be too close to the cursor colour, switch it to the cell's
background colour.

Xterm by default always just swaps a cell's foreground and background
colours to display the cursor. With the changes here, mintty effectively
does that as its fallback and so should do no worse than xterm.


git-svn-id: http://mintty.googlecode.com/svn/branches/1.1@1290 2782a51e-c42b-11dd-927d-974589228060